feat: check installed Apps
